Add lightweight auto-research kernel
