+++ loopx/capabilities/auto_research/__init__.py
@@ -48,6 +48,14 @@ from .core import (
     validate_research_evidence_event,
     validate_research_hypothesis,
 )
+from .kernel import (
+    LIGHTWEIGHT_AUTO_RESEARCH_EVIDENCE_SCHEMA_VERSION,
+    LIGHTWEIGHT_AUTO_RESEARCH_HYPOTHESIS_SCHEMA_VERSION,
+    LIGHTWEIGHT_AUTO_RESEARCH_RESULT_SCHEMA_VERSION,
+    lightweight_hypothesis,
+    run_builtin_lightweight_demo,
+    run_lightweight_auto_research,
+)
 
 __all__ = [
     "AUTO_RESEARCH_DEFAULT_GOAL_ID",
@@ -96,4 +104,10 @@ __all__ = [
     "validate_research_contract",
     "validate_research_evidence_event",
     "validate_research_hypothesis",
+    "LIGHTWEIGHT_AUTO_RESEARCH_EVIDENCE_SCHEMA_VERSION",
+    "LIGHTWEIGHT_AUTO_RESEARCH_HYPOTHESIS_SCHEMA_VERSION",
+    "LIGHTWEIGHT_AUTO_RESEARCH_RESULT_SCHEMA_VERSION",
+    "lightweight_hypothesis",
+    "run_builtin_lightweight_demo",
+    "run_lightweight_auto_research",
 ]
